Fix Dashboard bugs and add new Partner functionality
Dashboard would not update the partners and programs after logging in. Setting a useEffect to run on mount fixed the problem of them only being called by the App. Now when the user logs out and back in without reloading, it can properly fetch data from the database for the dashboards components.

Dashboard now dispatches calls to remove the partners and programs from state on a logout button click.

PartnerForm is now displayed on the dashboard with a view being available if the user clicks on a specific partner.